refactor(web): replace history API routes with server functions
Addresses review feedback on #584: the change-history data was exposed
via public /api/history-db/* GET routes. Replace them with "use server"
functions in lib/history-actions.ts (getHistoryIndex, getBuildChanges,
getEntityTimeline, getResourceIndex, getFileDiff, getStringChanges) that
query the history database directly and are invoked from the client
components via React Query — no public REST surface.

- remove the six /api/history-db/* route handlers
- remove lib/history-db.ts (server functions import historyDb from
  @jitaspace/db-history directly)
- drop the now-unused client fetchers + zod parsing from lib/history.ts
  and lib/resource-history.ts (schemas / types / metadata retained)

Also fixes a latent bug: app/api/history-db/build/[build]/route.ts was
never committed (caught by the `build` .gitignore rule), so that endpoint
would have 404'd in a fresh checkout — moot now it's a server function.
